int = float

+++ index.py
@@ -6,6 +6,6 @@ for i in range(50, 100, +3):
     print(i)
 
 A = input("welk getal vermenigvuldigen met 5? ")
-B = float(A)
+B = int(A)
 C = B*5
 print(C)
